+++ src/Client.php
@@ -3,9 +3,11 @@
 namespace Qless;
 
 use Qless\Exceptions\ExceptionInterface;
+use Qless\Exceptions\RuntimeException;
 use Qless\Exceptions\UnknownPropertyException;
 use Qless\Jobs\Collection as JobsCollection;
 use Qless\Subscribers\QlessCoreSubscriber;
+use Qless\Workers\Collection as WorkersCollection;
 use Redis;
 
 /**
@@ -20,11 +22,11 @@ use Redis;
  * @method string requeue(string $worker, string $queue, string $jid, string $klass, string $data, int $delay, ...$args)
  * @method string pop(string $queue, string $worker, int $count)
  * @method int length(string $queue)
- * @method int heartbeat(...$args)
+ * @method float heartbeat(...$args)
  * @method int retry(string $jid, string $queue, string $worker, int $delay, string $group, string $message)
  * @method int cancel(string $jid)
  * @method int unrecur(string $jid)
- * @method bool|string fail(string $jid, string $worker, string $group, string $message, string $data = null)
+ * @method bool|string fail(string $jid, string $worker, string $group, string $message, ?string $data = null)
  * @method string[] jobs(string $state, int $offset = 0, int $count = 25)
  * @method bool|string get(string $jid)
  * @method string multiget(string[] $jid)
@@ -39,7 +41,9 @@ use Redis;
  * @method string workers(?string $workerName = null)
  *
  * @property-read JobsCollection $jobs
+ * @property-read WorkersCollection $workers
  * @property-read Config $config
+ * @property-read Redis $redis
  * @property-read LuaScript $lua
  */
 class Client implements EventsManagerAwareInterface
@@ -55,6 +59,9 @@ class Client implements EventsManagerAwareInterface
     /** @var JobsCollection */
     private $jobs;
 
+    /** @var WorkersCollection */
+    private $workers;
+
     /** @var Redis */
     private $redis;
 
@@ -67,6 +74,9 @@ class Client implements EventsManagerAwareInterface
     /** @var float */
     private $redisTimeout = 0.0;
 
+    /** @var int */
+    private $redisDatabase = 0;
+
     /** @var string */
     private $workerName;
 
@@ -76,12 +86,15 @@ class Client implements EventsManagerAwareInterface
      * @param string $host    Can be a host, or the path to a unix domain socket.
      * @param int    $port    The redis port [optional].
      * @param float  $timeout Value in seconds (optional, default is 0.0 meaning unlimited).
+     * @param int   $database Redis database (optional, default is 0).
      */
-    public function __construct(string $host = '127.0.0.1', int $port = 6379, float $timeout = 0.0)
+    public function __construct(string $host = '127.0.0.1', int $port = 6379, float $timeout = 0.0, int $database = 0)
     {
         $this->redisHost = $host;
         $this->redisPort = $port;
         $this->redisTimeout = $timeout;
+        $this->redisDatabase = $database;
+
         $this->workerName = gethostname() . '-' . getmypid();
 
         $this->redis = new Redis();
@@ -91,9 +104,8 @@ class Client implements EventsManagerAwareInterface
 
         $this->lua = new LuaScript($this->redis);
         $this->config = new Config($this);
-
         $this->jobs = new JobsCollection($this);
-        $this->jobs->setEventsManager($this->getEventsManager());
+        $this->workers = new WorkersCollection($this);
     }
 
     /**
@@ -116,7 +128,11 @@ class Client implements EventsManagerAwareInterface
     {
         return new QlessCoreSubscriber(
             function (Redis $redis) {
-                $redis->connect($this->redisHost, $this->redisPort, $this->redisTimeout);
+                if ($redis->connect($this->redisHost, $this->redisPort, $this->redisTimeout)) {
+                    if ($this->redisDatabase !== 0) {
+                        $this->redis->select($this->redisDatabase);
+                    }
+                }
             },
             $channels
         );
@@ -125,8 +141,8 @@ class Client implements EventsManagerAwareInterface
     /**
      * Call a specific q-less command.
      *
-     * @param string $command
-     * @param mixed ...$arguments
+     * @param  string $command
+     * @param  mixed  ...$arguments
      * @return mixed|null
      *
      * @throws ExceptionInterface
@@ -142,8 +158,8 @@ class Client implements EventsManagerAwareInterface
     /**
      * Call a specific q-less command.
      *
-     * @param string $command
-     * @param array $arguments
+     * @param  string $command
+     * @param  array $arguments
      * @return mixed
      *
      * @throws ExceptionInterface
@@ -168,10 +184,9 @@ class Client implements EventsManagerAwareInterface
     {
         switch ($name) {
             case 'jobs':
-                $collection = $this->jobs;
-                $collection->setEventsManager($this->getEventsManager());
-
-                return $collection;
+                return $this->jobs;
+            case 'workers':
+                return $this->workers;
             case 'config':
                 return $this->config;
             case 'lua':
@@ -188,7 +203,7 @@ class Client implements EventsManagerAwareInterface
      *
      * @return void
      */
-    public function flush()
+    public function flush(): void
     {
         $this->redis->flushDB();
     }
@@ -198,7 +213,7 @@ class Client implements EventsManagerAwareInterface
      *
      * @return void
      */
-    public function reconnect()
+    public function reconnect(): void
     {
         $this->redis->close();
         $this->connect();
@@ -208,9 +223,17 @@ class Client implements EventsManagerAwareInterface
      * Perform connection to the Redis server.
      *
      * @return void
+     *
+     * @throws RuntimeException
      */
-    private function connect()
+    private function connect(): void
     {
-        $this->redis->connect($this->redisHost, $this->redisPort, $this->redisTimeout);
+        if ($this->redis->connect($this->redisHost, $this->redisPort, $this->redisTimeout) == false) {
+            throw new RuntimeException('Can not connect to the Redis server.');
+        }
+
+        if ($this->redisDatabase !== 0 && $this->redis->select($this->redisDatabase) == false) {
+            throw new RuntimeException('Can not select the Redis database.');
+        }
     }
 }